Extract __msvc_xlocinfo_types.hpp.
